requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.5.0
blinker==1.4
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.4
greenlet==3.1.1
h11==0.13.0
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.3.6
MarkupSafe==3.0.2
more-itertools==8.10.0
oauthlib==3.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.3.0
pymongo==4.10.1
pyparsing==2.4.7
python-dotenv==1.0.1
PyYAML==6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.15.0
wadllib==1.3.6
wsproto==1.0.0 
